deprecate Color class. Added optional includeAlpha to Color.toHex.
Added IRGB, modified IRGBA to extend IRGB.
Added RGB, and RGBA classes, to replace Color class.

Color class will be repurposed in the future to be an abstraction of the RGB/RGBA class,
to allow easily switching "modes".

As a stopgap, Color.toHex now accepts an optional boolean parameter, includeAlpha, to produce
a hex format with the alpha byte included, in the form of 0xRRGGBBAA. This uses RGBA class internally.